frontend view + s3 issue

+++ app/Http/Repository/FilesHandler.php
@@ -1,51 +1,75 @@
 <?php
 
 namespace App\Http\Repository;
-use App\Models\Files;
 use File;
-
+use Illuminate\Support\Facades\Storage;
+use App\Models\{ Files , Folder};
 
 class FilesHandler{
 
     public function uploadMedia($request)
     {
-        $folderId = $request->folder_id;
-        $fileList = [];
-        $videoExtension = ['mp4' , 'webm'];
-        // dd("inside file");
-        if($request->hasFile('files'))
-        {
-            foreach($request->file('files') as $file){
-                
-                $fileName = $file->getClientOriginalName();
-                $extension = $file->extension();
-                $name = time()."-".$fileName;
-                $file->move(public_path('uploads'), $name);
-                $type =in_array($extension , $videoExtension) ? 1 : 2;
-                $fileList[] = ['folder_id' => $folderId , 'type' => $type , 'path' => $name , 'extension' => $extension];
-            }
-            
-            Files::insert($fileList);    
-        }
+        try{
+            $folderId = $request->folder_id;
+            $fileList = [];
+            $videoExtension = ['mp4' , 'webm'];
+            // dd("inside file");
 
-        return ["success" => true , "msg" => "Files Added Successfully"];
+            $folder = Folder::find($folderId);
+
+            if($request->hasFile('files'))
+            {
+                foreach($request->file('files') as $file){
+
+                    $fileName = $file->getClientOriginalName();
+                    $extension = $file->extension();
+                    $name = time() . "-" . $fileName;
+                    Storage::disk('s3')->put($folder->name.'/'.$name, file_get_contents($file));     
+
+                    // $fileName = $file->getClientOriginalName();
+                    // $extension = $file->extension();
+                    // $name = time()."-".$fileName;
+                    // $file->move(public_path('uploads'), $name);
+                    $type =in_array($extension , $videoExtension) ? 1 : 2;
+                    $fileList[] = ['folder_id' => $folderId , 'type' => $type , 'path' => $name , 'extension' => $extension];
+                }
+                
+                Files::insert($fileList);    
+            }
+    
+            return ["success" => true , "msg" => "Files Added Successfully"];
+
+        }catch(\Exception $e){
+            dd("s3 is not working" , $e->getMessage());
+            return response()->json(['success' => false , "msg" => $e->getMessage()]);
+        }
     }
 
     public function deleteMedia($request)
     {
-        $fileId = $request->file_id;
 
-        $file = Files::where('id' , $fileId)->first();
-        $fileName = $file->path;
-        // dd(public_path('uploads')."/".$fileName);
-        if(File::exists(public_path('uploads')."/".$fileName))
-        {
-            File::delete(public_path('upload')."/".$fileName);
+        try{
+            $fileId = $request->file_id;
+    
+            $file = Files::where('id' , $fileId)->first();
+            $folder = Folder::find($file->folder_id);
+            $fileName = $folder->name.'/'.$file->path;
+    
+            //file path in s3
+            if(Storage::disk('s3')->exists($fileName))
+            {
+                Storage::disk('s3')->delete($fileName);
+                $file->delete();
+                return ["success" => true , "msg" => "File Deleted Successfully"];
+    
+            }else{
+                return ["success" => false , "msg" => "File Not Found"];
+            }
+
+        }catch(\Exception $e){
+            return ["success" => false , "msg" => $e->getMessage()];
         }
 
-        $file->delete();
-
-        return ["success" => true , "msg" => "File Deleted Successfully"];
 
     }
